+++ src/components/Card/index.jsx
@@ -1,0 +1,4 @@
+import Card from './Basic';
+import DisplayCard from './DisplayCard';
+
+export { Card, DisplayCard };
