+++ src/components/Input/index.jsx
@@ -1,0 +1,4 @@
+import Input from './Basic';
+import UnderlinedInput from './Underlined';
+
+export { Input, UnderlinedInput };
