--- src/components/common/index.jsx
@@ -1,6 +1,0 @@
-import Modal from './Modal';
-import Button from './Button';
-import Card from './Card';
-import Input from './Input';
-
-export { Modal, Button, Card, Input };
