+++ src/components/complex/CardOwnerNameInput.jsx
@@ -1,6 +1,6 @@
 import styled from 'styled-components';
 
-import { Input } from '../common';
+import { Input } from '..';
 
 import CARD_RULE from '../../constants';
 import { CardInfoContext } from '../../contexts';
