+++ src/components/complex/CardPasswordInput.jsx
@@ -1,6 +1,6 @@
 import styled from 'styled-components';
 
-import { Input } from '../common';
+import { Input } from '..';
 
 import { CardInfoContext } from '../../contexts';
