+++ src/components/complex/CardValidDateInput.jsx
@@ -1,4 +1,4 @@
-import { Input } from '../common';
+import { Input } from '..';
 
 import CARD_RULE from '../../constants';
 import { CardInfoContext } from '../../contexts';
