+++ src/components/complex/index.jsx
@@ -3,7 +3,6 @@ import CardValidDateInput from './CardValidDateInput';
 import CardOwnerNameInput from './CardOwnerNameInput';
 import CardCVCInput from './CardCVC';
 import CardPasswordInput from './CardPasswordInput';
-import DisplayCard from './DisplayCard';
 
 export {
   CardNumberInput,
@@ -11,5 +10,4 @@ export {
   CardOwnerNameInput,
   CardCVCInput,
   CardPasswordInput,
-  DisplayCard,
 };
