+++ src/components/index.jsx
@@ -1,0 +1,6 @@
+import { Input, UnderlinedInput } from './Input';
+import Button from './Button';
+import { Card } from './Card';
+import Modal from './Modal';
+
+export { Input, UnderlinedInput, Button, Card, Modal };
