+++ src/pages/CardListPage/index.jsx
@@ -2,7 +2,7 @@ import { memo } from 'react';
 import { useNavigate } from 'react-router-dom';
 import styled from 'styled-components';
 
-import Cards from '../../components/Cards';
+import Cards from './Cards';
 
 const StyledPage = styled.div`
   background: #fff;
